+++ src/lib/profile.ts
@@ -1,15 +1,7 @@
 import type { PrismaClient } from '@prisma/client';
-import { z } from 'zod';
+import type { z } from 'zod';
 import { APIError } from './errors';
-
-export const profileSchema = z.object({
-  goals: z.string().optional(),
-  imageS3ObjectKey: z.string().nullish(),
-});
-export type ProfileSchema = typeof profileSchema;
-
-export const profilePartialSchema = profileSchema.partial();
-export type ProfilePartialSchema = typeof profilePartialSchema;
+import type { ProfilePartialSchema } from './zodSchemas';
 
 export class ProfileRepo {
   constructor(private readonly prisma: PrismaClient) {}
