+++ src/routes/profile/+page.server.ts
@@ -3,9 +3,9 @@ import { ExerciseEventRepo } from '$lib/exerciseEvent';
 import { JournalEntryRepo } from '$lib/journalEntry';
 import { MetricRepo } from '$lib/metric';
 import { prisma } from '$lib/prisma';
-import { ProfileRepo, profileSchema } from '$lib/profile';
+import { ProfileRepo } from '$lib/profile';
 import { getSessionOrRedirect } from '$lib/utils';
-import { fileUploadSchema } from '$lib/zodSchemas';
+import { fileUploadSchema, profileSchema } from '$lib/zodSchemas';
 import { fail } from '@sveltejs/kit';
 import sharp from 'sharp';
 import { setError, superValidate } from 'sveltekit-superforms/server';
